+++ src/main/java/com/davistiba/wedemyserver/repository/CartRepository.java
@@ -6,12 +6,14 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+
 public interface CartRepository extends CrudRepository<Cart, Integer> {
 
     @Modifying
     @Transactional
     @Query(value = "INSERT INTO cart(created_at, course_id, user_id, price) VALUES (UTC_TIMESTAMP(), ?, ?, ?)", nativeQuery = true)
-    Integer addToCartCustom(Integer courseId, Integer userId, Double price);
+    Integer addToCartCustom(Integer courseId, Integer userId, BigDecimal price);
 
     @Modifying
     @Transactional
